aptiv: linux: add vae2is and vale2is ops to fix build failure
Issue: LIN1023-7148

Now the KVM configure is enabled by default, but when using the following macro:
 #define __tlbi(op, ...) __tlbi_##op(__VA_ARGS__)

which is controlled by CONFIG_NXP_S32CC_ERRATUM_ERR050481, to define the tlbi
instructions, there will be following error:

  arch/arm64/include/asm/tlbflush.h:80:25: error: implicit declaration of function '__tlbi_vae2is' [-Werror=implicit-function-declaration]
  |    80 | #define __tlbi(op, ...) __tlbi_##op(__VA_ARGS__)
  |       |                         ^~~~~~~

This is because there is lack of definitions of __tlbi_vae2is and
__tlbi_vale2is under CONFIG_NXP_S32CC_ERRATUM_ERR050481, so define them
to fix the above error.

The new added definitions refer to section "2.81 ERR050481" of the ERRATA
document "S32G2_0P77B.pdf", which is from vendor's website.

Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>